+++ lib/raster3d/long.c
@@ -49,7 +49,7 @@ void Rast3d_long_decode(unsigned char *source, long *dst, int nofNums,
     dest += nofNums - 1;
     while (source != srcStop) {
         *dest = *source--;
-        if ((eltLength >= RASTER3D_LONG_LENGTH) && (*dest != 0))
+        if ((eltLength >= (int)RASTER3D_LONG_LENGTH) && (*dest != 0))
             Rast3d_fatal_error("Rast3d_long_decode: decoded long too long");
         dest--;
     }
@@ -61,7 +61,7 @@ void Rast3d_long_decode(unsigned char *source, long *dst, int nofNums,
         while (source != srcStop) {
             *dest *= 256;
             *dest += *source--;
-            if ((eltLength >= RASTER3D_LONG_LENGTH) && (*dest != 0))
+            if ((eltLength >= (int)RASTER3D_LONG_LENGTH) && (*dest != 0))
                 Rast3d_fatal_error("Rast3d_long_decode: decoded long too long");
             dest--;
         }
