+++ raster/r.fill.stats/main.c
@@ -155,7 +155,7 @@ long int estimate_mem_needed(long int cols, char *mode)
 void print_weights_matrix(long int rows, long int cols)
 {
     int i, j;
-    int weight_matrix_line_length = 80;
+    size_t weight_matrix_line_length = 80;
     char weight_matrix_line_buf[weight_matrix_line_length + 1];
     char weight_matrix_weight_buf[weight_matrix_line_length + 1];
